+++ src/index.js
@@ -141,7 +141,6 @@ import StoresApi from './api/StoresApi'
 import TitlesApi from './api/TitlesApi'
 import UsersApi from './api/UsersApi'
 import VouchersApi from './api/VouchersApi'
-import OAuth from './OAuth'
 
 /**
 * ERROR_UNKNOWN.<br>
@@ -929,11 +928,5 @@ export {
     * The VouchersApi service constructor.
     * @property {module:api/VouchersApi}
     */
-    VouchersApi,
-
-    /**
-    * The OAuth object constructor.
-    * @property {module:api/OAuth}
-    */
-    OAuth
+    VouchersApi
 }