+++ test/api/VouchersApi.spec.js
@@ -17,17 +17,18 @@
 
 import expect from 'expect.js'
 import Occ from '../config'
+import {requestAccessToken, clearAccessToken} from '../utils'
 
 let instance
 
 beforeEach((done) => {
     instance = new Occ.VouchersApi()
-    Occ.ApiClient.instance.requestAccessToken(Occ.authentication)
+    requestAccessToken(Occ.ApiClient.instance, Occ.authentication)
         .then(done)
 })
 
 afterEach(() => {
-    Occ.ApiClient.instance.clearAccessToken()
+    clearAccessToken(Occ.ApiClient.instance)
 })
 
 describe('VouchersApi', () => {